feat: OrInterfaceTrait is added
